Discuss the average image and location heatmap
Although, as it turns out, the average image appears fairly random.

Additionally, the "location" information which should be contained within the
pickle file appears to be missing at present.

Further work on producing a heatmap of sign locations in the original images
will have to wait until this is resolved.

This references https://github.com/udacity/sdc-issue-reports/issues/123