Use better name for logits, and switch to AdamOptimizer
Previously, it appears that the optimizer I had chosen was minimizing
accuracy.  This was, of course, not the intended result, but there was
no visible bug that would possibly lead to this behavior.

Accordingly, switch to a different optimizer rather than
GradientDescent, to see if that made a difference.

At the same time, it's extremely likely that the neural net used in this
model is way too simplistic to be able to learn to correctly classify 43
different street signs.

It seems like that, instead, a 768 -> 500 -> 200 -> 43 Neural Net, since
it is deeper, would have a better ability to abstract out and perform
well on the dataset.

That, along with DropOut and ReLU, will be the next architecture that I
try on this dataset.