Fix learning rate decay bug
The `global_steps` tf.Variable stores the number of evaluations of
the `cost` tensor.  This tensor is evaluated once per batch.  Hence,
for batch sizes greater than 1, this tensor is evaluated a factor of
`batch_size` less frequently than the number of training images.

That is, while there are ~39,000 training images, if the `batch_size`
is, say, 50, then `global_steps` will only reach ~780 per epoch.

Accordingly, reduce 50,000 for `decay_steps` to 1,000.